factor out components to reduce size of GardenPage
